test save commit

+++ backend/controllers/cards.js
@@ -1,42 +1,72 @@
-const Card = require('../models/card');
+const Card = require('../models/card.js');
+const CastError = require('../errors/CastError');
+const ForbiddenError = require('../errors/ForbiddenError');
+const NotFoundError = require('../errors/NotFoundError');
 
-function getCards(req, res) {
+function getCards(req, res, next) {
   Card.find()
-    .then((card) => res.send(card))
-    .catch(() => {
-      res.status(500).send({ message: 'Ошибка на сервере' });
-    });
+    .then((card) => res.status(200).send(card))
+    .catch(next);
 }
 
-function createCard(req, res) {
-  Card.create({ owner: req.user, ...req.body })
-    .then((card) => res.send(card))
-    .catch((err) => {
-      if (err.name === 'ValidationError') {
-        return res.status(400).send({ message: 'Данные невалидны' });
-      }
-      return res.status(500).send({ message: 'Ошибка на сервере' });
-    });
+function createCard(req, res, next) {
+  const { name, link } = req.body;
+
+  Card.create({
+    name,
+    link,
+    owner: req.user.id,
+    createdAt: Date.now(),
+  })
+    .then((card) => res.status(200).send(card))
+    .catch(next);
 }
 
-function deleteCard(req, res) {
+function deleteCard(req, res, next) {
   const { cardId } = req.params;
-  Card.findByIdAndRemove(cardId)
-    .orFail(new Error('notExistId'))
-    .then((card) => res.send(card))
-    .catch((err) => {
-      if (err.message === 'notExistId') {
-        res.status(404).send({ message: 'Нет карточки с таким id' });
-      } else if (err.name === 'CastError') {
-        res.status(400).send({ message: 'Неверный id' });
+  Card.findById(cardId)
+    .orFail(new NotFoundError('Запрашиваемый ресурс не найден'))
+    .then((card) => {
+      if (!card) throw new CastError('Неверный id карточки');
+      if (card.owner.toString() === req.user.id.toString()) {
+        Card.findByIdAndRemove(cardId)
+          .then((cardToRemove) => res.status(200).send(cardToRemove))
+          .catch(next);
       } else {
-        res.status(500).send({ message: 'Ошибка на сервере' });
+        throw new ForbiddenError('Вы не можете удалять карточки, которые были созданы не вами');
       }
-    });
+    })
+    .catch(next);
+}
+
+function addLike(req, res, next) {
+  const { cardId } = req.params;
+  Card.findByIdAndUpdate(
+    cardId,
+    { $addToSet: { likes: req.user.id } },
+    { new: true },
+  )
+    .orFail(new NotFoundError('Запрашиваемый ресурс не найден'))
+    .then((card) => res.status(200).send(card))
+    .catch(next);
+}
+
+function removeLike(req, res, next) {
+  const { cardId } = req.params;
+  Card.findByIdAndUpdate(
+    cardId,
+    { $pull: { likes: req.user.id } },
+    { new: true },
+  )
+    .orFail(new NotFoundError('Запрашиваемый ресурс не найден'))
+    .then((card) => res.status(200).send(card))
+    .catch(next);
 }
 
 module.exports = {
   getCards,
   createCard,
   deleteCard,
+  addLike,
+  removeLike,
 };
